Changes objects.count to report the maximum label number
As discussed in the issue/3582 thread, currently there is a
heavy reliance in the CP code that there are no missing objects
in the labelmasks. Thus objects.count is often used to initialize
vectors of length 1:objects.count where each entry location
is supposed to correpond to a label (so label=10 should be at position 10).
This breaks, however, in cases where there are missing objects
in the label mask, as it can happen e.g. with externaly created masks
or IdentifySecondary.
Changing the objects.count would be a rather dirty way to fix this,
but essentially would work.