Adds a `maxlabel` property to the objects
This would be an alternative way to tackle to problem
of missing objects discussed in the issue: give the objects a property that
corresponds to the maximum labels.

Note that for this to properly work, `objects.count` would need
to be replaced in a couple of modules.

Also measurements should then be initialized with np.nan instead of
np.zeros